+++ RELEASENOTES.md
@@ -7,7 +7,7 @@
 * AKS and Arc Container Images:
   - feat: Configmap update for CCP (v2 + v1 schema support) (https://github.com/Azure/prometheus-collector/pull/1056)
 * Pipeline/Docs/Templates Updates:
-  - 
+  - fix: move filesystemwatcher to hash based golang appoach for windows (https://github.com/Azure/prometheus-collector/pull/1144)
 
 
 ## Release 04-15-2025